test(Steps): update test cases
